+++ linear_match.h
@@ -8,7 +8,7 @@ double sqr(double x);
 double dist(double x0, double y0, double x1, double y1);
 
 namespace LinearMatch {
-	double matchArrs(int*, int*, int*);
+	double matchArrs(double*, double*, double*);
 };
 
 #endif
